+++ src/main/kotlin/kr/hhplus/be/server/domain/concert/eventListener/SelloutRankingOnReservationCancelledListener.kt
@@ -10,7 +10,7 @@ import org.springframework.transaction.event.TransactionPhase
 import org.springframework.transaction.event.TransactionalEventListener
 
 @Component
-class ReservationCancelledEventListener(
+class SelloutRankingOnReservationCancelledListener(
     private val selloutRankingService: SelloutRankingService,
     private val errorHandler: EventErrorHandler
 ) {
