+++ src/main/kotlin/kr/hhplus/be/server/domain/reservation/eventListener/DataPlatformReservationCreatedListener.kt
@@ -1,6 +1,6 @@
-package kr.hhplus.be.server.domain.payment.eventListener
+package kr.hhplus.be.server.domain.reservation.eventListener
 
-import kr.hhplus.be.server.domain.payment.event.PaymentCompletedEvent
+import kr.hhplus.be.server.domain.reservation.event.ReservationCreatedEvent
 import kr.hhplus.be.server.global.client.ConcertDataPlatformClient
 import kr.hhplus.be.server.global.event.EventErrorHandler
 import kr.hhplus.be.server.global.event.EventErrorHandling
@@ -10,26 +10,26 @@ import org.springframework.scheduling.annotation.Async
 import org.springframework.stereotype.Component
 
 @Component
-class PaymentCompletedEventListener(
+class DataPlatformReservationCreatedListener(
     private val concertDataPlatformClient: ConcertDataPlatformClient,
     private val errorHandler: EventErrorHandler
 ) {
-
+    
     private val logger = KotlinLogging.logger {}
-
+    
     @EventErrorHandling(sendToDLQ = false, critical = false)
     @Async
     @EventListener
-    fun handle(event: PaymentCompletedEvent) {
-        errorHandler.handleEventSafely(event, "PaymentCompletedEvent") {
-            concertDataPlatformClient.sendPaymentData(
-                paymentId = event.paymentId,
-                userId = event.userId,
+    fun handle(event: ReservationCreatedEvent) {
+        errorHandler.handleEventSafely(event, "ReservationCreatedEvent") {
+            concertDataPlatformClient.sendReservationData(
                 reservationId = event.reservationId,
-                amount = event.amount,
-                operationType = "PAYMENT_COMPLETED"
+                userId = event.userId,
+                concertId = event.concertId,
+                seatId = event.seatId,
+                operationType = "RESERVATION_CREATED"
             )
-            logger.info { "데이터 플랫폼 결제 완료 정보 전송 완료: ${event.paymentId}" }
+            logger.info { "데이터 플랫폼 예약 생성 정보 전송 완료: ${event.reservationId}" }
         }
     }
 }
