+++ src/main/kotlin/kr/hhplus/be/server/domain/reservation/infrastructure/ReservationJpaRepository.kt
@@ -4,7 +4,11 @@ import kr.hhplus.be.server.domain.reservation.models.Reservation
 import org.springframework.data.domain.Page
 import org.springframework.data.domain.Pageable
 import org.springframework.data.jpa.repository.JpaRepository
+import org.springframework.data.jpa.repository.Lock
+import org.springframework.data.jpa.repository.Query
+import org.springframework.data.repository.query.Param
 import org.springframework.stereotype.Repository
+import jakarta.persistence.LockModeType
 import java.time.LocalDateTime
 
 @Repository
